Porting Networking additions from Java SDK; Batch deletion support across top level collections; *GetBy returns null (#3038)
* Added Martin's changes

* porting Networking additions from Java SDK

* AvailabilitySet.VirtualMachineIds now return ISet<string> instead of IList<string>
(port from Java SDK)

* porting batch deletion support to top level resource collections from Java

* GetBy* should return null for NotFound, just like in Java SDK, not throw.

* new/modified app gateway tests

+++ src/ResourceManagement/Network/Microsoft.Azure.Management.Network.Fluent/Domain/INetworkUsage.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 namespace Microsoft.Azure.Management.Network.Fluent
 {
-    using Models;
+    using Microsoft.Azure.Management.Network.Fluent.Models;
     using Microsoft.Azure.Management.ResourceManager.Fluent.Core;
 
     /// <summary>
@@ -20,7 +20,7 @@ namespace Microsoft.Azure.Management.Network.Fluent
         /// Gets the maximum count of the resources that can be allocated in the
         /// subscription.
         /// </summary>
-        int Limit { get; }
+        long Limit { get; }
 
         /// <summary>
         /// Gets the name of the type of usage.
@@ -30,6 +30,6 @@ namespace Microsoft.Azure.Management.Network.Fluent
         /// <summary>
         /// Gets the current count of the allocated resources in the subscription.
         /// </summary>
-        int CurrentValue { get; }
+        long CurrentValue { get; }
     }
 }
